add more computations
